Log to standard error rather than standard output
Due to stdio buffering, logging to the latter isn't very
useful (unless stdout is a tty, which for a daemon is a
marginal use case; and even then it's customary to log to
stderr).

(A concrete data point: running with '-F' under systemd, the
first batch of log messages got to the journal after 4
days.)

When at it, direct also the startup error messages to stderr.

(Leave -h and -v alone, printing to stdout seems to be
widespread there, although it's a mixed bag (double-checked
with a few BSD and GNU utilities).)